+++ make/copy_vcrt.lua
@@ -1,3 +1,6 @@
-local output = ...
+local output, arch = ...
 local fs = require 'bee.filesystem'
-require 'msvc'.copy_vcrt('x64', fs.current_path() / output)
+require 'msvc'.copy_vcrt(
+    arch == "x86" and 'x86' or 'x64',
+    fs.current_path() / output
+)