+++ script/core/completion/completion.lua
@@ -1064,14 +1064,14 @@ local function tryLabelInString(label, source)
     if not source or source.type ~= 'string' then
         return label
     end
-    local str = parser.grammar(label, 'String')
-    if not str then
+    local state = parser.parse(label, 'String')
+    if not state or not state.ast then
         return label
     end
-    if not matchKey(source[1], str[1]) then
+    if not matchKey(source[1], state.ast[1]) then
         return nil
     end
-    return util.viewString(str[1], source[2])
+    return util.viewString(state.ast[1], source[2])
 end
 
 local function mergeEnums(a, b, source)
@@ -1354,20 +1354,6 @@ local function getCallEnumsAndFuncs(source, index, oop, call)
         end
     end
     if source.type == 'doc.type.function' then
-        --[[
-        always use literal index, that is:
-        ```
-        ---@class Class
-        ---@field f(x: number, y: boolean)
-        local c
-
-        c.f(1, true) -- correct
-        c:f(1, true) -- also correct
-        ```
-        --]]
-        if oop then
-            index = index - 1
-        end
         local arg = source.args[index]
         if arg and arg.extends then
             return pushCallEnumsAndFuncs(vm.getDefs(arg.extends))
@@ -1959,6 +1945,7 @@ local function makeCache(uri, position, results)
     cache.position= position
     cache.word    = word:lower()
     cache.length  = #word
+    cache.uri     = uri
 end
 
 local function isValidCache(word, result)
@@ -1981,6 +1968,9 @@ end
 local function getCache(uri, position)
     local cache = workspace.getCache 'completion'
     if not cache.results then
+        return nil
+    end
+    if cache.uri ~= uri then
         return nil
     end
     local text  = files.getText(uri)
@@ -2058,6 +2048,9 @@ local function completion(uri, position, triggerCharacter)
     await.delay()
     tracy.ZoneBeginN 'completion #1'
     local state = files.getState(uri)
+    if not state then
+        return nil
+    end
     results = {}
     clearStack()
     tracy.ZoneEnd()