+++ test/crossfile/definition.lua
@@ -867,3 +867,93 @@ print(t.<?x?>)
         ]]
     }
 }
+
+local originRuntimePath = config.get 'Lua.runtime.path'
+
+config.set('Lua.runtime.path', {
+    './?.lua'
+})
+TEST {
+    {
+        path = 'a.lua',
+        content = [[
+return {
+    <!x!> = 1,
+}
+]],
+    },
+    {
+        path = 'b.lua',
+        content = [[
+local t = require 'a'
+print(t.<?x?>)
+        ]]
+    }
+}
+
+config.set('Lua.runtime.path', {
+    '/home/?.lua'
+})
+TEST {
+    {
+        path = '/home/a.lua',
+        content = [[
+return {
+    <!x!> = 1,
+}
+]],
+    },
+    {
+        path = 'b.lua',
+        content = [[
+local t = require 'a'
+print(t.<?x?>)
+        ]]
+    }
+}
+
+config.set('Lua.runtime.pathStrict', true)
+config.set('Lua.runtime.path', {
+    './?.lua'
+})
+TEST {
+    {
+        path = 'a.lua',
+        content = [[
+return {
+    <!x!> = 1,
+}
+]],
+    },
+    {
+        path = 'b.lua',
+        content = [[
+local t = require 'a'
+print(t.<?x?>)
+        ]]
+    }
+}
+
+config.set('Lua.runtime.path', {
+    '/home/?.lua'
+})
+TEST {
+    {
+        path = '/home/a.lua',
+        content = [[
+return {
+    <!x!> = 1,
+}
+]],
+    },
+    {
+        path = 'b.lua',
+        content = [[
+local t = require 'a'
+print(t.<?x?>)
+        ]]
+    }
+}
+
+config.set('Lua.runtime.pathStrict', false)
+config.set('Lua.runtime.path', originRuntimePath)